extract LoginForm component
